requirements.txt
Pillow~=9.4.0
ipython~=8.11.0
pip~=23.0.1
tornado~=6.2
pyzmq~=25.0.1
pexpect~=4.8.0
comm~=0.1.2
traitlets~=5.9.0
parso~=0.8.3
jedi~=0.18.2
pytz~=2022.7.1
setuptools~=65.5.0
scipy~=1.10.1
joblib~=1.2.0
psutil~=5.9.4
pandas~=1.5.3
python-dateutil~=2.8.2
Pygments~=2.14.0
decorator~=5.1.1
packaging~=23.0
backcall~=0.2.0
pickleshare~=0.7.5
wcwidth~=0.2.6
debugpy~=1.6.6
ptyprocess~=0.7.0
seaborn~=0.12.2
cycler~=0.11.0
scikit-learn~=1.2.2
threadpoolctl~=3.1.0
six~=1.16.0
imblearn~=0.0
asttokens~=2.2.1
contourpy~=1.0.7
executing~=1.2.0
fonttools~=4.39.0
ipykernel~=6.21.3
pyparsing~=3.0.9
kiwisolver~=1.4.4
platformdirs~=3.1.1